Adds listings query parameters
Sending a get request to /listings with parameters for game and system
will now show all listings where any of the lookingFor games match
the query. If no query params are provided, all listings will be returned.

This functionality will ultimately be used to match a user's listings up
with those from other users based on what game a user is trying to trade.